# [FLINK-19710] Revert implementation of PerfContext back to __thread to avoid performance regression