--- a/monitoring/perf_context_imp.h
+++ b/monitoring/perf_context_imp.h
@@ -16,7 +16,7 @@ extern PerfContext perf_context;
 extern __thread PerfContext perf_context_;
 #define perf_context (*get_perf_context())
 #else
-extern thread_local PerfContext perf_context;
+extern __thread PerfContext perf_context;
 #endif
 #endif