RocksDB introduce performance regression via [replaced __thread with thread_local keyword](https://github.com/facebook/rocksdb/commit/d6ec288703c8fc53b54be9e3e3f3ffd6a7487c63).
This PR reverts `PerfContext` back to `__thread` and abort some usage as it would not be used Flink.