@carp84 2020-10-26
@Myasuka could you attach the performance data before / after this change / fix? Thanks