@Myasuka 2020-10-26
Below is the experimental result after apply this fix on a linux machine with 500GB memory in total.

| Benchmark                                                   			| original RocksDB-6.11 throughput | Fixed RocksDB-6.11 throughput 				|
|-----------------------------------------------------------------------|----------------------------------|--------------------------------------------|
| org.apache.flink.state.benchmark.ListStateBenchmark.listAdd 			|489.340911	± 5.614968	ops/ms 	   |	507.793520 ± 7.887317  ops/ms  		|
| org.apache.flink.state.benchmark.ListStateBenchmark.listAddAll 		|289.259124	± 4.757700	 ops/ms	   |	299.358497 ± 10.935809  ops/ms  		|
| org.apache.flink.state.benchmark.ListStateBenchmark.listAppend 		|460.467266	± 6.567233	ops/ms 	   |	483.354214 ± 6.141290  ops/ms  		|
| org.apache.flink.state.benchmark.ListStateBenchmark.listGet	 		|122.195119	± 0.583669	ops/ms 	   |	138.067541 ± 0.519469  ops/ms  		|
| org.apache.flink.state.benchmark.ListStateBenchmark.listGetAndIterate |122.376738	± 0.705641	ops/ms 	   |	137.312238 ± 1.323462  ops/ms  		|
| org.apache.flink.state.benchmark.ListStateBenchmark.listUpdate 		|495.459603	± 4.561395	ops/ms 	   |	509.520106 ± 5.077400  ops/ms  		|
| org.apache.flink.state.benchmark.MapStateBenchmark.mapAdd 			|431.088394	± 20.468059	 ops/ms	   |	446.252946 ± 22.396690  ops/ms  		|
| org.apache.flink.state.benchmark.MapStateBenchmark.mapContains 		| 48.730236	± 0.457412	ops/ms 	   |	54.539764  ± 0.118094  ops/ms 			|
| org.apache.flink.state.benchmark.MapStateBenchmark.mapEntries 		|329.641797	± 7.774533	ops/ms 	   |	347.769279 ± 8.160462  ops/ms  		|
| org.apache.flink.state.benchmark.MapStateBenchmark.mapGet 			| 49.164114	± 0.177340	ops/ms 	   |	53.877028  ± 0.265754  ops/ms 			|
| org.apache.flink.state.benchmark.MapStateBenchmark.mapIsEmpty 		| 42.816571	± 0.139690	ops/ms 	   |	46.796282  ± 0.214785  ops/ms 			|
| org.apache.flink.state.benchmark.MapStateBenchmark.mapIterator 		|331.323677	± 7.935423	ops/ms 	   |	351.181264 ± 8.571055  ops/ms  		|
| org.apache.flink.state.benchmark.MapStateBenchmark.mapKeys	 		|321.080059	± 12.947762	ops/ms 	   |	349.881785 ± 8.268457  ops/ms  		|
| org.apache.flink.state.benchmark.MapStateBenchmark.mapPutAll 			|100.037343	± 6.779264	ops/ms 	   |	94.216012  ± 7.537000  ops/ms 			|
| org.apache.flink.state.benchmark.MapStateBenchmark.mapRemove 			|426.177218	± 29.927329	 ops/ms	   |	445.132551 ± 28.183995  ops/ms  		|
| org.apache.flink.state.benchmark.MapStateBenchmark.mapUpdate 			|428.597067	± 21.329073	 ops/ms	   |	446.231137 ± 18.235515  ops/ms  		|
| org.apache.flink.state.benchmark.MapStateBenchmark.mapValues	 		|330.840970	± 7.683108	ops/ms 	   |	349.290260 ± 6.695283  ops/ms  		|
| org.apache.flink.state.benchmark.ValueStateBenchmark.valueAdd 		|426.132608	± 22.066362	 ops/ms	   |	436.521025 ± 31.498314  ops/ms  		|
| org.apache.flink.state.benchmark.ValueStateBenchmark.valueGet 		|620.187399	± 44.498021	 ops/ms	   |	670.520548 ± 30.156394  ops/ms  		|
| org.apache.flink.state.benchmark.ValueStateBenchmark.valueUpdate 		|427.154900	± 27.734762	 ops/ms	   |	434.678702 ± 31.288398  ops/ms  		|