@carp84 2020-10-27
Thanks for the follow up @Myasuka , from the numbers we could see a stable performance gain for almost all cases of our state micro benchmark.