@Myasuka 2020-10-27
The key performance regression is introduced by usage of `thread-local`, and this PR mainly revert the usage of `thread-local` back to `__thread` and remove related usage.
Apart from existing tests, I also run it with my program to check the statistics like block cache hit/miss count could still get as expected with this PR.