@azagrebin 2020-10-28
Thanks for opening the PR @Myasuka 
As I understand the main problem is the replacement of `__thread` with `thread_local`.
Could you write more details about why we need the other code removal?
If it is not going to be used in Flink why does it hurt? or is it also somehow related to the performance regression?