@Myasuka 2020-10-29
@azagrebin 
In a nutshell, we need to revert `thread-local` back to `__thread`. However, initializer for thread-local variable must be a constant expression (refer to [gcc-tls](https://gcc.gnu.org/onlinedocs/gcc/Thread-Local.html)), that's why I had to drop all those constructors and destructor. That's why I need to drop tests add in https://github.com/facebook/rocksdb/pull/4919 and add default initializer for those values in the struct of `PerfContext`

I also leave the `PerfContextByLevel` there to not drop them.